Add files via upload
This project involves the creation of a Data Visualization Software using Java programming language and the NetBeans IDE. The software's primary function is to facilitate the visualization of data through various chart types, including bar charts, 3D bar charts, line charts, and pie charts. The data will be presented in graphical forms, providing a clear representation of information trends. A notable feature of the software is its ability to save the generated charts in PNG file format. This ensures that users can preserve and share their visualizations for presentations, reports, or further analysis. The software will allow users to effortlessly export the charts as image files, enhancing the utility and versatility of the tool.